--- components/t.html
@@ -1,16 +1,0 @@
-<div
-  class="MuiBox-root css-1ixbp0l"
-  style="max-width: 100%; height: 180px; overflow: hidden"
->
-  <iframe
-    srcdoc=''
-    frameborder="0"
-    style="
-      width: 1400px;
-      height: 700px;
-      border: 0px;
-      transform: scale(0.3);
-      transform-origin: 0px 0px;
-    "
-  ></iframe>
-</div>
